% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_process.R
\name{filter_process}
\alias{filter_process}
\title{Multivariate filter}
\usage{
filter_process(x, filter, ...)
}
\arguments{
\item{x}{a multivariate time series provided as a q by T matrix}

\item{filter}{an n by q by l array}

\item{...}{further parameters passed to the underlying c routine,
currently without functionality}
}
\value{
a matrix of dimension n by T where
  y[t] = f[,,1] * x[t] + f[,,2] * x[t-1] ...
}
\description{
Multivariate filter
}
