% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredmd.R
\docType{data}
\name{fredmd}
\alias{fredmd}
\title{FRED Monthly Data for Macroeconomic Research}
\source{
<https://research.stlouisfed.org/econ/mccracken/fred-databases/>.
}
\usage{
data(fredmd)
}
\description{
Dataset of US macroeconomic variables obtained from the FRED database. It has been
transformed according to the procedure of McCracken and Ng (2016) (appendix, p.26).
Moreover, the first and last two time observations have been removed as well as
series with missing values in between.
}
\references{
Michael W. McCracken and Serena Ng (2016). FRED-MD: A Monthly Database
for Macroeconomic Research. Journal of Business & Economic Statistics, 34(4),
574–589. https://doi.org/10.1080/07350015.2015.1086655
}
