% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_filter.R
\name{multivariate_filter}
\alias{multivariate_filter}
\title{Multivariate filter}
\usage{
multivariate_filter(
  x,
  filter_coefficients,
  lags = seq_len(dim(filter_coefficients)[3])
)
}
\arguments{
\item{x}{A multivariate time series provided either as a matrix (rows correspond
to cross-sectional units and columnts to observations in the time domain) or a
multivariate object of class `ts` or `zoo`.}

\item{filter_coefficients}{An n by q by l array holding
filter coefficients.}

\item{lags}{A vector of length dim(filter_coefficients)[3]
specifying the lag index.}
}
\value{
A matrix of dimension n by T - max(c(0, lags)) - max(c(0, -lags)) where
n and T are the input dimension, and

  y[t] = f[,,1] * x[t - lags[1]] + f[,,2] * x[t-lags[2]] ...
}
\description{
Applies linear filter to a multivariate time series.
}
