% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.spca}
\alias{plot.spca}
\title{Diagnostics plot for an \code{spca} object.}
\usage{
\method{plot}{spca}(x, ...)
}
\arguments{
\item{x}{An object of type \code{spca}.}

\item{...}{Further pass-through arguments.}
}
\description{
Diagnostics plot for an \code{spca} object.
}
