% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.dpca}
\alias{print.dpca}
\title{S3 methods for s3 classes "dpca" and "spca"
Print method for object of class \code{dpca}.}
\usage{
\method{print}{dpca}(x, ...)
}
\arguments{
\item{x}{An object of type \code{dpca}.}

\item{...}{Further pass-through arguments.}
}
\description{
S3 methods for s3 classes "dpca" and "spca"
Print method for object of class \code{dpca}.
}
