% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.spca}
\alias{print.spca}
\title{Print method for object of class \code{spca}.}
\usage{
\method{print}{spca}(x, ...)
}
\arguments{
\item{x}{An object of type \code{spca}.}

\item{...}{Further pass-through arguments.}
}
\description{
Print method for object of class \code{spca}.
}
