% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.summary.dpca}
\alias{print.summary.dpca}
\title{Summary method for object of class \code{summary.dpca}.}
\usage{
\method{print}{summary.dpca}(x, ...)
}
\arguments{
\item{x}{An object of type \code{summary.dpca}.}

\item{...}{Further pass-through arguments.}
}
\description{
Summary method for object of class \code{summary.dpca}.
}
