% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.summary.spca}
\alias{print.summary.spca}
\title{Summary method for object of class \code{summary.spca}.}
\usage{
\method{print}{summary.spca}(x, ...)
}
\arguments{
\item{x}{An object of type \code{summary.spca}.}

\item{...}{Further pass-through arguments.}
}
\description{
Summary method for object of class \code{summary.spca}.
}
