% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_filter.R
\name{recursive_filter}
\alias{recursive_filter}
\title{Multivariate recursive filter.}
\usage{
recursive_filter(x, filter_coefficients)
}
\arguments{
\item{x}{A multivariate time series provided either as a matrix (rows correspond
to cross-sectional units and columnts to observations in the time domain) or a
multivariate object of class \code{\link[stats]{ts}} or \code{\link[zoo]{zoo}}.}

\item{filter_coefficients}{An \eqn{n} by \eqn{m} by \eqn{p} array holding
filter coefficients, where \eqn{p} is the maximum lag order.}
}
\value{
A matrix of dimension \eqn{n} by \eqn{T - p} where
\eqn{n} and \eqn{T} are the input dimension, and

\deqn{%
  y_t = f_{(,,1)} * y_{t-1} + ... +  f_{(,,p)} * y_{t-p} + x_t
}{%
  y[t] = f[,,1] * y[t-1] + ... +  f[,,p] * y[t-p] + x[t]
}
}
\description{
Apply a recursive filter to a multivariate time series.
}
