% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_r.R
\name{select_r}
\alias{select_r}
\title{Select number of static factors.}
\usage{
select_r(
  x,
  penalty = c("IC1", "IC2", "IC3"),
  penalty_scales = seq(0, 2, by = 0.01),
  n_path,
  max_r,
  ...
)
}
\arguments{
\item{x}{A data frame of variables (each row one time observation)
or a data matrix (n by T).}
}
\description{
Select the number of "static" factors for a static factor sequence
via the criteria of Bai and Ng (2002) but using the stability interval
method suggested by Hallin and Liska (2007.
}
