% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{summary.dpca}
\alias{summary.dpca}
\title{Summary method for object of class \code{dpca}.}
\usage{
\method{summary}{dpca}(object, ...)
}
\arguments{
\item{object}{An object of type \code{spca}.}

\item{...}{Further pass-through arguments.}
}
\description{
Summary method for object of class \code{dpca}.
}
