% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{summary.spca}
\alias{summary.spca}
\title{Summary method for object of class \code{spca}.}
\usage{
\method{summary}{spca}(object, ...)
}
\arguments{
\item{object}{An object of type \code{spca}.}

\item{...}{Further pass-through arguments.}
}
\description{
Summary method for object of class \code{spca}.
}
