% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_filter.R
\name{multivariate_filter}
\alias{multivariate_filter}
\title{Multivariate filter}
\usage{
multivariate_filter(
  x,
  filter_coefficients,
  lags = seq_len(dim(filter_coefficients)[3])
)
}
\arguments{
\item{x}{A multivariate time series provided as a q by T matrix.}

\item{filter_coefficients}{An n by q by l array holding
filter coefficients.}

\item{lags}{A vector of length dim(filter_coefficients)[3]
specifying the lag index.}
}
\value{
A matrix of dimension n by T where
  y[t] = f[,,1] * x[t - lags[1]] + f[,,2] * x[t-lags[2]] ...
}
\description{
Applies linear filter to a multivariate time series.
}
