% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{plot.dpca}
\alias{plot.dpca}
\title{Diagnostics plot for a \code{dpca} object.}
\usage{
\method{plot}{dpca}(x, ...)
}
\arguments{
\item{x}{An object of type \code{dpca}.}

\item{...}{Further pass-through arguments.}
}
\description{
Diagnostics plot for a \code{dpca} object.
}
