% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_filter.R
\name{recursive_filter}
\alias{recursive_filter}
\title{Multivariate recursive filter.}
\usage{
recursive_filter(x, filter_coefficients)
}
\arguments{
\item{x}{A multivariate time series provided as a q by T matrix.}

\item{filter_coefficients}{An n by q by p array holding
filter coefficients, where p is the maximum lag order.}
}
\description{
Apply a recursive filter to a multivariate time series.
}
