% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_filter.R
\name{recursive_filter}
\alias{recursive_filter}
\title{Multivariate recursive filter.}
\usage{
recursive_filter(x, filter_coefficients)
}
\arguments{
\item{x}{A multivariate time series provided either as a matrix (rows correspond
to cross-sectional units and columnts to observations in the time domain) or a
multivariate object of class `ts` or `zoo`.}

\item{filter_coefficients}{An n by q by p array holding
filter coefficients, where p is the maximum lag order.}
}
\value{
A matrix of dimension n by T - p where
n and T are the input dimension, and

  y[t] = f[,,1] * y[t-1] + ... +  f[,,p] * y[t-p] + x[t]
}
\description{
Apply a recursive filter to a multivariate time series.
}
