% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{spca}
\alias{spca}
\title{Static Principal Component Analysis and Large Dynamic Factor Model Estimation}
\usage{
spca(
  x,
  r,
  rsel = FALSE,
  rsel_crit = c("IC1", "IC2", "IC3"),
  n_path = floor(seq(nrow(x)/2, nrow(x), nrow(x)/20)),
  penalty_scales = seq(0, 2, by = 0.01)
)
}
\arguments{
\item{x}{Input data supplied either as a matrix (rows correspond to cross-sectional
units and columnts to observations in the time domain) or a multivariate object of
class `ts` or `zoo`.}

\item{r}{Number of static factors. If rsel = TRUE, this is
the maximum number of static factors.
#'}

\item{rsel}{Logical, if TRUE one of the Hallin & Liska criteria are used to
choose q from the data.}

\item{rsel_crit}{Criterion to select the number of factors using the
Hallin & Liska (2007, JASA) method. Either "IC1" or "IC2".}

\item{n_path}{Integer vector specifying which (nested) subsets of the
cross section are used in the Hallin & Liska procedure.}

\item{penalty_scales}{Tuning values for the penalty scaling parameter
c over which the q-path is optimized to stability.}
}
\value{
An object of class "spca".
}
\description{
Static Principal Component Analysis and Large Dynamic Factor Model Estimation
}
