% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{spca}
\alias{spca}
\title{Static Principal Component Analysis and Large Dynamic Factor Model Estimation}
\usage{
spca(
  x,
  r,
  rsel = FALSE,
  rsel_crit = c("IC1", "IC2", "IC3"),
  n_path = NULL,
  penalty_scales = seq(0, 2, by = 0.01)
)
}
\arguments{
\item{x}{Input data supplied either as a matrix (rows correspond to cross-sectional
units and columnts to observations in the time domain) or a multivariate object of
class \code{\link[stats]{ts}} or \code{\link[zoo]{zoo}}.}

\item{r}{Number of static factors. If \code{rsel} is set to \code{TRUE}, this is
the maximum number of static factors.}

\item{rsel}{Logical, if \code{TRUE} one of the Hallin & Liska criteria are used to
choose \code{r} from the data.}

\item{rsel_crit}{Criterion to select the number of factors using the
Hallin & Liska (2007, JASA) method. Either \code{"IC1"} or \code{"IC2"}.}

\item{n_path}{Integer vector specifying which (nested) subsets of the
cross section are used in the Hallin & Liska procedure. If unspecified,
a regular sequence of length \code{20} from \code{n/2} to \code{n} is used.}

\item{penalty_scales}{Tuning values for the penalty scaling parameter
\eqn{c} over which the \code{q}-path is optimized to stability.}
}
\value{
An object of class "spca", wrapping a list with the entries
\itemize{
  \item \code{xmean}: a vector holding the mean of each cross-sectional unit
  \item \code{cov}: variance-covariance-matrix of \code{x}
  \item \code{eig}: eigen decomposition of \code{cov}
  \item \code{factors}: an \eqn{r} times \eqn{T} dimensional matrix with the computed factors
  \item \code{cc}: (static) common component
  \item \code{ic}: (static) idiosyncratic component
  \item \code{HL_select}: results of the selection methodology of Hallin & Liska (2007),
}
see also \code{\link{select_r}}.
}
\description{
Static Principal Component Analysis and Large Dynamic Factor Model Estimation
}
